Refactored Film into its own module
